Verificación de token y permisos (Auth0)

+++ API/Routes/salesRoute.js
@@ -1,12 +1,22 @@
 const { Router } = require("express");
 const cors = require("cors");
+const {
+    checkRequiredPermissions,
+    validateAccessToken} = require("../Auth0/auth0.middleware.js");
+  const {
+  
+  } = require("../Auth0/auth0.permissions.js");
 const { getAllSales, createSale, getSaleById, updateSale, deleteSale } = require('../Controllers/sales.js')
 const { validateNewSale } = require("../Validators/sale.js");
 const {validate} = require("../Helpers/validateHelper.js")
 
 saleRouter = Router();
 
-saleRouter.get("/sale", cors(), async (req, res) => {
+saleRouter.get(
+    "/sale",
+    validateAccessToken,
+    checkRequiredPermissions([]),
+    async (req, res) => {
     //Si no hay productos en la BD, devuelve un arreglo vacío. NO es un error...
     try {
       const allSales = await getAllSales();
@@ -16,10 +26,14 @@ saleRouter.get("/sale", cors(), async (req, res) => {
     };
   });
 
-saleRouter.post("/sale", cors(),validate(validateNewSale), createSale);
+saleRouter.post("/sale", cors(), validateAccessToken, checkRequiredPermissions([]), validate(validateNewSale), createSale);
   //Si algún dato no es válido o falta, se lanzan los errores correspondientes. Faltan las funciones validadoras.
 
-saleRouter.get("/sale/:id", cors(), async (req, res) => {
+saleRouter.get(
+    "/sale/:id",
+    validateAccessToken,
+    checkRequiredPermissions([]),
+    async (req, res) => {
     try {
         const { id } = req.params;
         const sale = await getSaleById(id);
@@ -29,7 +43,11 @@ saleRouter.get("/sale/:id", cors(), async (req, res) => {
     };
 });
 
-saleRouter.put("/sale/:id", cors(), async (req, res) => {
+saleRouter.put(
+    "/sale/:id",
+    validateAccessToken,
+    checkRequiredPermissions([]),
+    async (req, res) => {
     try {
         const { id } = req.params;
         const update = req.body; //Hay que preguntar a los del frontend si están usando Axios o similares...
@@ -41,7 +59,11 @@ saleRouter.put("/sale/:id", cors(), async (req, res) => {
     };
 });
 
-saleRouter.delete("/sale/:id", cors(), async (req, res) => {
+saleRouter.delete(
+    "/sale/:id",
+    validateAccessToken,
+    checkRequiredPermissions([]),
+    async (req, res) => {
     try {
         const { id } = req.params;
         await deleteSale(id);
